fix(common matchers): accept arrays as well as strings
We have some shorthand functions to add matchers for labels like
`alertname`, `cluster` and so on. These are currently typed to only
accept strings. The [upstream configuration][config-docs] show that
these can accept multiple values when given as a regular expression with
alternates.

That's what our existing `matchArrayOrString` helper does. It's just not
used here - let's do that.

Now the code

```jsonnet
inhibit_rule.sourceMatcher.alert('KubeNodeNotReady')
+ inhibit_rule.targetMatcher.alert(['PromScrapeFlapping', 'PromScrapeFailed'])
+ inhibit_rule.targetMatcher.namespace(['nsone', 'nstwo']
```

Will produce the expected regular expression for the `alertname` label,
as it already did for the `namespace` label.

The same argument applies to `cluster`, `team` and `severity`, so we
update those here too.

[config-docs]: https://prometheus.io/docs/alerting/latest/configuration/#matcher